[EP/graph_trainer] Add PaddedExpertParallel and _local_scalar_dense fallback
Add PaddedExpertParallel, a graph-compilation-compatible alternative to
ExpertParallel that avoids data-dependent ops (_local_scalar_dense).

Standard ExpertParallel uses dynamic split sizes for all-to-all and
dynamic permutation indices, both of which introduce _local_scalar_dense
ops that Inductor cannot compile. PaddedExpertParallel instead:
1. Pads routed_input so each rank sends/receives equal-sized chunks
2. Uses equal-split all-to-all (split_sizes=None)
3. Replaces dynamic _permute with static reshape+transpose

Also updates the graph_trainer DeepSeek V3 parallelize to use
PaddedExpertParallel when EP is enabled, force-balance routing on
all MoE modules, and annotate PaddedExpertParallel for FX tracing.

This is temporary infrastructure that will be removed when non-even
routing support lands in the compiler.

stack-info: PR: https://github.com/pytorch/torchtitan/pull/2703, branch: xmfan/stack/24

+++ torchtitan/models/common/moe/moe.py
@@ -63,7 +63,10 @@ def _run_experts_grouped_mm(
     x: torch.Tensor,
     num_tokens_per_expert: torch.Tensor,
 ) -> torch.Tensor:
-    offsets = torch.cumsum(num_tokens_per_expert, dim=0, dtype=torch.int32)
+    # Two-step cumsum to work around Inductor lowering bug where
+    # cumsum(..., dtype=torch.int32) produces int64 in post-grad passes.
+    offsets = torch.cumsum(num_tokens_per_expert.to(torch.int32), dim=0)
+    offsets = offsets.to(torch.int32)
 
     h = F.silu(
         torch._grouped_mm(x.bfloat16(), w1.bfloat16().transpose(-2, -1), offs=offsets)
